pyproject.toml
[project]
name = "allumfurniture"
version = "0.1.0"
description = ""
authors = [
    {name = "Semon Horochih",email = "HorSemAlex@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = false

dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
